Add basic search for the hostel list
Added a basic search of the hostel list to show some expected
functionality of the final application.

'setSearchParams()' is putting search params '?filter=' in the URL and
rerendering the router. Then the value of the input is set to what is
in the filter search param and the list of hostels is based on the
the filter search param.

At this time, the filtering is only finding hostels that start with
the searchparams. In future, this will be changed to filter based on
what the name and description of the hostel contains.